requirements.txt
langchain-mistralai==0.2.11
fastapi[standard]==0.116.1
gradio==5.38.0
openpyxl==3.1.5
xlrd>=2.0.1
pandas==2.3.0
SQLAlchemy==2.0.37
httpx==0.28.1
python-dotenv==1.0.0